Add coeff to grid size approx, change rect algo
Subtract two from the contour count before sqrt'ing it to get the grid
size guess. The text elements below the grid dilate up to be around the
size of the grid elements, so at low actual grid sizes they throw the
guess up to e.g. 5 instead of 4.

Additionally change the grid border bounding box filtering algorithm to
be more relaxed, now targeting specifically shapes that are not
isosceles trapezoids, instead of targeting shapes that are not squares.
This gives us leeway when the user hasn't yet aligned the camera to be
parallel with the screen.