Fix grid size being wrong when zooming in ...
... on small grid sizes. Apparently when the user has zoomed in so that
the only thing visible in the area of interest is the grid itself and
the grid size is small, our little contour rounding coefficient magic
rounds 4 down to 3.

Fixed by checking if the result of the square root is an integer and
only applying the modifier if not. That is, if the approx grid size is
exactly some grid size, use that instead of applying the coefficient.